torchattacks/requirements.txt
numpy==1.22.4
Pillow==9.1.1
torch==1.9.1
torchvision==0.10.1
typing_extensions==4.2.0
